requirements.txt
apache-airflow-providers-microsoft-mssql==1.0.1
apache-airflow-providers-odbc==1.0.1
apache-airflow-providers-apache-spark==3.0.0
pymssql==2.2.1